ssh_key_streamlocalbindmask.go
package ssh_config

// StreamLocalBindMask
//         Sets the octal file creation mode mask (umask) used when creat‐
//         ing a Unix-domain socket file for local or remote port forward‐
//         ing.  This option is only used for port forwarding to a Unix-
//         domain socket file.
//
//         The default value is 0177, which creates a Unix-domain socket
//         file that is readable and writable only by the owner.  Note
//         that not all operating systems honor the file mode on Unix-
//         domain socket files.
//

func init() {
	funcInit := func() (res string) {
		// TODO
		panic("Not implemented")
		return
	}

	funcValid := func(value string) (res bool) {
		// TODO
		panic("Not implemented")
		return
	}

	funcParse := func(input string) (values []string, err error) {
		// TODO
		panic("Not implemented")
		return
	}

	sshInit(StreamLocalBindMask, funcInit)
	sshValid(StreamLocalBindMask, funcValid)
	sshParse(StreamLocalBindMask, funcParse)
}
